+++ grpc/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name    = "protobuffers"
+name    = "grpc"
 version = "0.1.0"
 edition = "2021"
 
@@ -19,6 +19,8 @@ tonic      = "0.7"
 prost      = "0.10"
 env_logger = "0.10.0"
 log        = { version = "0.4.20", features = [ "max_level_info" ] }
+fred       = "6.3.1"
+redis      = { path = "../redis" }
 
 
 [build-dependencies]
