--- src/consts.rs
@@ -1,4 +1,0 @@
-pub const SUBMISSION_KEY: &str = "submission_id";
-pub const SUBMISSIONS_LIST: &str = "submissions";
-pub const SCRAPPING_SET: &str = "currently_scrapping";
-pub const DEFAULT_START_SUBMISSION_ID: u32 = 1223456789;
